Tehtävät 4.21

+++ osa4/controllers/blogs.js
@@ -53,6 +53,11 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
     try {
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+        if (!request.token || !decodedToken.id) {
+            return response.status(401).
+                json({error: 'token missing or invalid'})
+        }
         const res = await Blog.findByIdAndRemove(request.params.id)
         if (res) {
             response.status(200).end()
@@ -60,7 +65,15 @@ blogsRouter.delete('/:id', async (request, response) => {
             response.status(404).end()
         }
     } catch (exception) {
-        response.status(500).json(exception)
+        if (exception.name === 'JsonWebTokenError') {
+            response.status(401).json({error: exception})
+        } else {
+            console.log(exception)
+            response.status(500).json({
+                error: 'Something went horribly wrong!',
+                exception: exception,
+            })
+        }
     }
 })
